requirements.txt
joblib==1.0.0
numpy==1.19.1
opencv-python==4.3.0.36
openslide-python==1.1.2
pytorch==1.9.1
sklearn==0.0
tensorboard==2.4.1
tensorboardx==2.1
torchvision==0.10.1